Add forward AD for linalg.det and simplify its backward
This PR is in preparation for implementing `logdet` and `slogdet` as
structured kernels + implementing them with more efficient derivatives

We implement forward AD for det. We also simplify the implementation of
the backward, and leave a note on how to implement it properly for
singular matrices. We leave thad for future work.

Note (by looking at the OpInfo) that the current implementation passes
the same tests as the one before. We skip the forward-over-backward in
the singular case, as that one was not working in the gradgrad case
either.

ghstack-source-id: 476c91da1bb3f272b6ccd4a58f80e73cbdb86ac5
Pull Request resolved: https://github.com/pytorch/pytorch/pull/79487